move microseconds_sliding_window
