+++ ydb/core/persqueue/common/ut/ya.make
@@ -1,0 +1,11 @@
+UNITTEST_FOR(ydb/core/persqueue/common)
+
+SRCS(
+    microseconds_sliding_window_ut.cpp
+)
+
+PEERDIR(
+    library/cpp/testing/unittest
+)
+
+END()
