+++ ydb/core/persqueue/ut/ya.make
@@ -47,7 +47,6 @@ SRCS(
     user_info_ut.cpp
     pqrb_describes_ut.cpp
     partition_scale_manager_graph_cmp_ut.cpp
-    microseconds_sliding_window_ut.cpp
     fetch_request_ut.cpp
     utils_ut.cpp
     list_all_topics_ut.cpp
