+++ ydb/core/persqueue/ya.make
@@ -21,7 +21,6 @@ SRCS(
     user_info.cpp
     write_meta.cpp
     write_quoter.cpp
-    microseconds_sliding_window.cpp
     dread_cache_service/caching_service.cpp
     write_id.cpp
     tracing_support.cpp
